# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TODO List API
  description: |
    A comprehensive RESTful API for managing tasks organized within multiple lists.

    ## Features
    - Create and manage multiple todo lists
    - Add, update, and delete tasks within lists
    - Set deadlines and priorities for tasks
    - Track completion status
    - Advanced querying and filtering
    - Task statistics and productivity insights

    ## Authentication
    Currently, this API does not require authentication. All endpoints are publicly accessible.

    ## Rate Limiting
    API requests are rate-limited to 1000 requests per hour per IP address.

    ## Data Validation
    All request bodies are validated against JSON schemas. Invalid requests will return 400 Bad Request with validation details.
  version: 1.0.0
  contact:
    name: Development Team
    email: dev@todoapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.todolist.example.com
    description: Production server
security: []
...
paths:
...
  /api/tasks:
    get:
      tags: [Queries]
      summary: Query tasks with filtering
      description: Get tasks with advanced filtering and sorting options
      operationId: queryTasks
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [completed, pending]
          description: Filter by completion status
        - in: query
          name: priority
          schema:
            type: string
            enum: [low, medium, high, urgent]
          description: Filter by priority level
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [title, priority, deadline, createdAt, updatedAt]
          description: Field to sort by
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort order
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of results
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of results to skip
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTaskResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/tasks/due-this-week:
    get:
      tags: [Queries]
      summary: Get tasks due this week
      description: Retrieve all tasks that are due within the current week
      operationId: getTasksDueThisWeek
      responses:
        '200':
          description: Tasks due this week retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskResponse'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Tasks due this week retrieved successfully'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/tasks/overdue:
    get:
      tags: [Queries]
      summary: Get overdue tasks
      description: Retrieve all tasks that are past their deadline
      operationId: getOverdueTasks
      responses:
        '200':
          description: Overdue tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskResponse'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Overdue tasks retrieved successfully'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/tasks/statistics:
    get:
      tags: [Queries]
      summary: Get task statistics
      description: Get overall statistics about all tasks
      operationId: getTaskStatistics
      responses:
        '200':
          description: Task statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TaskStatistics'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Task statistics retrieved successfully'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/tasks/grouped:
    get:
      tags: [Queries]
      summary: Get tasks grouped by status
      description: Get tasks organized by their completion status
      operationId: getTasksGrouped
      responses:
        '200':
          description: Grouped tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GroupedTasks'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Grouped tasks retrieved successfully'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/insights/productivity:
    get:
      tags: [Queries]
      summary: Get productivity insights
      description: Get productivity analytics and insights
      operationId: getProductivityInsights
      responses:
        '200':
          description: Productivity insights retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ProductivityInsights'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Productivity insights retrieved successfully'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
...
  schemas:
...
    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
          example: '987fcdeb-51a2-43d7-8f9e-123456789abc'
        listId:
          type: string
          format: uuid
          description: ID of the list this task belongs to
          example: '123e4567-e89b-12d3-a456-426614174000'
        title:
          type: string
          description: Title of the task
          example: 'Complete project documentation'
        description:
          type: string
          nullable: true
          description: Optional detailed description of the task
          example: 'Write comprehensive documentation for the new feature'
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
        priority:
          type: string
          enum: [low, medium, high, urgent]
          description: Priority level of the task
          example: 'high'
        deadline:
          type: string
          format: date-time
          nullable: true
          description: Optional deadline for the task
          example: '2025-07-10T17:00:00.000Z'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the task was created
          example: '2025-07-01T11:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
          example: '2025-07-02T15:00:00.000Z'
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the task was completed
          example: null

    # Request Models
...
    TaskStatistics:
      type: object
      properties:
        totalTasks:
          type: integer
          description: Total number of tasks
          example: 25
        completedTasks:
          type: integer
          description: Number of completed tasks
          example: 15
        pendingTasks:
          type: integer
          description: Number of pending tasks
          example: 10
        overdueTasks:
          type: integer
          description: Number of overdue tasks
          example: 3
        tasksWithDeadlines:
          type: integer
          description: Number of tasks with deadlines
          example: 18
        completionRate:
          type: number
          format: float
          description: Task completion rate as a percentage
          example: 60.0
        priorityBreakdown:
          type: object
          properties:
            low:
              type: integer
              example: 5
            medium:
              type: integer
              example: 12
            high:
              type: integer
              example: 6
            urgent:
              type: integer
              example: 2
    GroupedTasks:
      type: object
      properties:
        completed:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
          description: List of completed tasks
        pending:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
          description: List of pending tasks
    ProductivityInsights:
      type: object
      properties:
        averageCompletionTime:
          type: number
          format: float
          description: Average time to complete tasks in hours
          example: 24.5
        mostProductiveDay:
          type: string
          description: Day of week with most completions
          example: 'Monday'
        tasksByPriority:
          type: object
          properties:
            low:
              type: integer
              example: 5
            medium:
              type: integer
              example: 12
            high:
              type: integer
              example: 6
            urgent:
              type: integer
              example: 2
        recentActivity:
          type: object
          properties:
            lastWeekCompletions:
              type: integer
              example: 8
            thisWeekCompletions:
              type: integer
              example: 5
            trend:
              type: string
              enum: [increasing, decreasing, stable]
              example: 'decreasing'

    # Pagination Models
    PaginatedTaskResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        pagination:
          type: object
          properties:
            total:
              type: integer
              description: Total number of items
              example: 100
            limit:
              type: integer
              description: Number of items per page
              example: 20
            offset:
              type: integer
              description: Number of items skipped
              example: 0
            hasNext:
              type: boolean
              description: Whether there are more items
              example: true
            hasPrev:
              type: boolean
              description: Whether there are previous items
              example: false
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 'Tasks retrieved successfully'
        timestamp:
          type: string
          format: date-time

    # Error Models
    ValidationError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: 'VALIDATION_ERROR'
            message:
              type: string
              example: 'Validation failed'
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: 'title'
                  message:
                    type: string
                    example: 'Title is required'
                  value:
                    oneOf:
                      - type: string
                      - type: number
                      - type: boolean
                      - type: 'null'
        timestamp:
          type: string
          format: date-time
...
    InternalServerError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: 'INTERNAL_SERVER_ERROR'
            message:
              type: string
              example: 'An unexpected error occurred'
            requestId:
              type: string
              example: 'req_123456789'
        timestamp:
          type: string
          format: date-time
  responses:
    ValidationError:
      description: Bad Request - Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerError'
...